fix surrogate/source id column names
Previously, I had used `id` as the surrogate key on the events table,
and then had to remap the `id` key in the event data to an alternate
name. This is confusing, and apparently using bare `id` columns (instead
of `thing_id` on the `things` table) is a bad idea.

The slides had a mistake in this remapping, so fix them to use
`event_id` and allow the `id` column to hold the events' `id`s.

Caught during pdxpug presentation.