Update for 2Q PGConf 2017
Doing this as a live demo in 20 minutes means dropping the size
comparisons and exploring the payload data type.

Also update the load time for the JSON vs JSONB data types, which seems
to be much better - that may be some combination of a new hard drive, a
new kernel, and a new version of PostgreSQL. Size on disk remained the
same, though.